Create Python program to combine foremost and scalpel
Develop a Python program that takes four command-line arguments:(1)The name of the tool (foremost or scalpel), (2)The input file path, (3)The output directory path and (4)The location of the configuration file.This Python program should run both scalpel and foremost on the specified input file (the 2nd command-line argument) using the configuration file (the 4th command-line argument). The output from both tools should be stored in the directory specified by the output directory path (the 3rd command-line argument).